11_Django_and_Docker/task_manager/authentication/api_view.py
from django.contrib.auth import models
from rest_framework import viewsets

from authentication import serializers


class UserViewSet(viewsets.ModelViewSet):
    queryset = models.User.objects.all()
    serializer_class = serializers.UserSerializer


class GroupViewSet(viewsets.ModelViewSet):
    queryset = models.Group.objects.all()
    serializer_class = serializers.GroupSerializer
